Improve query object searching (#1325)
* Rename QueryStore to RegistrationQueryStore

QueryStore is a very generic name and as we'll be adding a load of new
Query classes over the coming months this should reduce scope for
confusion.

* Add base Query class and put reusable logic there

* Introduce sentinel default values

The previous implementation of the statements query had the slight flaw
that the default values for the keyword query arguments were `nil`, and
`nil` is meaningful in ActiveRecord, which converts it to `null`.

Instead if we use the `:ignore` symbol to mark arguments as 'omitted',
it frees up `nil` for `null` queries.

This change also shifts the building of the scope to the initializer
and allows us to just keep the scope attr instead of one for each
argument.

Almost inadvertantly, this format combined with the `:ignore` pattern
brings back chaining, which might be useful as we progress.

* Make existing state scopes use with_state

* Move extract_conditions to a module

Also add some specs covering it in more depth

* Add specs covering :ignore behaviour

We're checking the SQL in some of these because we're looking for the
absence of a condition, and doing this via the returned values doesn't
confirm omission.

* Remove chaining and make where_ methods private

We don't need this for now, possibly useful to expose later