Add missing test coverage for admin dashboard (#1356)
* Add missing test coverage for admin dashboard

Make the feature specs consistent across the admin dashboard. We may want to
rethink this a bit in the future and move some of the views into view
components that are tested instead, but this at least makes it all consistent
for now.

* Attempt to fix flakey spec

The spec testing for previous funding is failing with seed `53300` in CI. Its
not clear why and I can't replicate it/bisect it locally, but I expect another
application is created that we aren't expecting.

Ensure we update _all_ of the users other applications when testing previously
funded behaviour.