Create a list of valid event types
There's only one in it so far, ApplicationAcceptance. The idea is that
if we can reliably identify the events by type we can show them with
more or less prominence (or not at all) depending on the context.

So, an application creation might be important from a user's point of
view, but less so from a course's.