Add api/v2/npq-enrolments.csv endpoint (#1400)
* Pluralise ApplicationCsvSerializer

As we are serializing multiple applications, not a single application.

Remove `call` pattern and update specs to be consistent with enrolments serializer.

* Add EnrolmentsCsvSerializer

Serialize enrolments (applications), mimicing the serializer from ECF.

* Add enrolments CSV endpoint for API v2

Add the GET `/api/v2/enrolments.csv` endpoint.

* Add enrolments CSV endpoint swagger docs for API v2

Add the API docs for the GET `api/v2/enrolments.csv` endpoint.

* Keep API schema versioning in the schema files

We seem to have introduced a second way to manage different schema shapes for
different versions of the API.

Update to keep the version logic consistently in the individual schema files.

* Only return accepted applications in enrolments CSV

We only want to surface accpeted applications in the enrolments
CSV as these are the ones that are equivalent to the
`ParticipantProfile` records in ECF.