Add `Participant::Defer` service and endpoints (#1429)
* Add Participant::Defer service

Add the service to move a participant to a deferred state. This is lifted from
ECF; instead of looking for the matching profile we find the matching
application.

* Add /participants/npq/:id/defer endpoint

Adds the endpoint to enable providers to defer the training status of a participant.

* Ensure application is accepted with_declaration in factory

We would only have declarations on accepted applications.