Add `Participant::Withdraw` service and endpoints (#1430)
* Add Participant::Withdraw service

Add the service to move a participant to a withdrawn state. This is lifted from
ECF; instead of looking for the matching profile we find the matching
application .

* Add /participants/npq/:id/withdraw endpoint

Adds the endpoint to enable providers to withdraw the training status of a participant.

* DRY up participant action services

Add `Participants::Action` abstract class that individual actions can inherit
from. Move specs into a shared context.

* DRY up participant action request specs

Move specs into a single shared context and pass in the service.

* Rescue malformed request body and return 400

When the request contains a malformed body (without data/attributes) we want to
return a 400 with an error message.

* Fix flakey participant action spec

This was flakey because the lead provider was occasionally being instantiated
and then selected in the application anyway (via `sample`). To avoid this we
ensure the application lead provider is created first (by referencing it in the
name of the other lead provider).

* Fix withdraw route typo