Add declarations endpoint (#1445)
* Add `DeclarationSerializer` for all API versions

Adds the `DeclarationSerializer` to replicate the ECF serialization for each
API version. Adds convenience methods to `Declaration` to keep the serializer
as simple as possible.

* Add `Declarations::Query` service

Add the `Declarations::Query` service to enable querying declarations. The
functionality is lifted directly from ECF and supports all API versions.

* Add `/api/:version/declarations` endpoint

Add the GET `/api/:version/declarations` endpoint, which calls the
`Declarations::Query` and returns via the `Declarations::Serializer`. The
functionality for filtering has been lifted from ECF.

* Update `/declarations` to `/participant-declarations`

We want to keep parity with ECF on the routes.