Add convenience methods to `Declaration`/`Statement`/`LeadProvider`
These are the convenience methods we need for finding output fee statements and
checking the state of declarations as part of the void/claw back service.

In future we're going to look at introducing a state machine to the
`Declaration` model, similar to that in `StatementItem` to manage the state
transitions.