Add `Declarations::Void` service
Ports across the void declaration logic from ECF. Confusingly the `/void`
action of the API is used for both voiding declarations and clawing them back
(the action is determined by the state of the declaration). In ECF we split out
the void and clawback services but I opted to combine them here (as the void
service calls the clawback service anyway and the additional service just makes
it more confusing than it already is, IMO).

The validation rule to ensure only a `paid` declaration can be clawed back
turned out to be redundant (as the service treats non-paid declarations as a
void intention anyway) so I didn't bring this across.

This also brings across the `StatementAttacher` from ECF, complete with the
validation to make sure there is a next output fee statement.