Add declaration endpoint (#1446)
* Add GET `/api/:version/declarations/:id` endpoint

Add the GET `/api/:version/declarations/:id` endpoint, which will return a
single declaration by ID for the current lead provider.

* Fix `ineligible_for_funding_reason` when not `duplicate`

There was a bug in the previous implementation; it should be returning
`state_reason` and not `reason`. We can't easily test this switch
statement yet as there is only one `state_reason` at the moment.