Treat `awaiting_clawback` and `clawed_back` as `Void#clawing_back?`
To avoid confusion in the error messages we send to providers, we want to
treat `awaiting_clawback` and `clawed_back` declarations as 'clawing back'
when going through the `Void` service.

Add additional error to catch when a lead provider attempts to clawback a
declaration that is not in a paid state; I don't think this error is great
but its what we have in ECF (although a bug meant we didn't surface it in
ECF).

Remove `not_voidable` error as this can no longer be reached.