Auto-generate request/response  examples
Previously the request/response examples were generated from
the response schemas; instead, we want to use the actual
responses as the example.

Removes the examples from the CI check as they are now dynamic
and will change randomly.

Originally I tried to make this less of a change by having an
option to load in the default response example and override
certain keys on a per-example basis, but loading in the rswag
example manually isn't straight forward. I think this will be
better in the long run anyway as it will ensure an accurate
example response. The only issue may be that the response
examples change on each deploy; this needs TBC.

Remove hard-coded examples.