Fix response examples for participant actions
In the participant action response examples we want to show
the correct `training_status`. In the response schema we have
`training_status` as `active`, which was being used for
resume/defer/withdraw.

To get a more accurate response example we have to define the
whole response manually for each action.

Update shared context to allow customising the response
example on a case by case basis (defaulting to the response
schema). Override the response example for participant
actions.