Add `/participant-declarations/:id/void` endpoints (#1460)
* Add convenience methods to `Declaration`/`Statement`/`LeadProvider`

These are the convenience methods we need for finding output fee statements and
checking the state of declarations as part of the void/claw back service.

In future we're going to look at introducing a state machine to the
`Declaration` model, similar to that in `StatementItem` to manage the state
transitions.

* Add `Declarations::Void` service

Ports across the void declaration logic from ECF. Confusingly the `/void`
action of the API is used for both voiding declarations and clawing them back
(the action is determined by the state of the declaration). In ECF we split out
the void and clawback services but I opted to combine them here (as the void
service calls the clawback service anyway and the additional service just makes
it more confusing than it already is, IMO).

The validation rule to ensure only a `paid` declaration can be clawed back
turned out to be redundant (as the service treats non-paid declarations as a
void intention anyway) so I didn't bring this across.

This also brings across the `StatementAttacher` from ECF, complete with the
validation to make sure there is a next output fee statement.

* Add `/participant-declarations/:id/void` endpoint

Add the `/void` endpoint for participant declarations; this will either void or
clawback the declaration as per the logic in ECF.

* Add request specs for voiding declarations

* DRY up request specs with generic shared contexts

We've been adding a shared context for each action, but this isn't really
necessary as the actions for update/create on resource are largely
identical and we just need to switch out the service being called and the
serializer being used.

* Update `next_output_fee_statements` spec to use factories

Instead of asserting on the SQL we set up some test data.

* Treat `awaiting_clawback` and `clawed_back` as `Void#clawing_back?`

To avoid confusion in the error messages we send to providers, we want to
treat `awaiting_clawback` and `clawed_back` declarations as 'clawing back'
when going through the `Void` service.

Add additional error to catch when a lead provider attempts to clawback a
declaration that is not in a paid state; I don't think this error is great
but its what we have in ECF (although a bug meant we didn't surface it in
ECF).

Remove `not_voidable` error as this can no longer be reached.