Add declarations CSV endpoint (#1450)
* Add `DeclarationsCsvSerializer`

Add `DeclarationsCsvSerializer`; as per ECF this will return the same data as
the JSON serializer/API endpoint.

Unlike the other serializers in ECF this one will return an empty CSV if there
are no declarations (other serializers return the header rows).

* Add `api/:version/participant-declarations.csv` endpoint

Add the CSV endpoint for participant declarations in v1 and v2 of the API, as per ECF.

* Add swagger docs for declarations CSV endpoints

Add swagger documentation for the declarations CSV endpoints.

* DRY up the swagger enums/examples

We can reference the enums directly here; this will keep the documentation
inline with the code automatically as well.

* DRY up and fix swagger versioning

Previously we were duplicating some swagger schemas instead of building off
previous versions. When I cleaned this up I also noticed that the hashes were
passing by reference between the versions, so we need to `deep_dup` them prior
to modifying, or all versions would end up with the changes.

* Ensure view is forwarded to serializer

I had a test for this but it was passing by chance as I was testing for
`v1` and had hard-coded it also. Fix to pass the view through to the
serializer and add a test case to cover it.