Add Swagger docs for participant actions (#1437)
* Add swagger docs for participant action endpoints

Add the swagger documentation for the participant action endpoints
(resume/defer/withdraw).

* DRY up application documentation specs

We were creating specific support files for each action; we can instead have a
single generic spec file that can be customised to cover any POST request
against a resource. The PUT request for change-funded-place can use the
existing generic update shared context.

Minor updates to descriptions/summaries for participants to bring inline with
applications.

* Fix response examples for participant actions

In the participant action response examples we want to show
the correct `training_status`. In the response schema we have
`training_status` as `active`, which was being used for
resume/defer/withdraw.

To get a more accurate response example we have to define the
whole response manually for each action.

Update shared context to allow customising the response
example on a case by case basis (defaulting to the response
schema). Override the response example for participant
actions.

* Update response examples for accept/reject application

If we use the default response example here the `status` will
always be `accepted`. Instead, we have to manually override
the response example and adjust it for each action.

* Extract base examples from swagger docs

To prevent hard-coding the base examples prior to overriding certain
attributes in the API doc specs we can extract the default example from
the schema and use it as a starting point.

* Fix required request attributes in docs

When specifying if an attribute is required you need to specify both
`required: true` on the attribute and `required: %i[attribute_name]` on
the parent.

* Fix v3 participant response examples

As the base example contains `deferral` and `withdrawl` examples we just
need to keep the applicable one for the response (and neither for
`resume`).

* Fix withdrawal typos