Add `ParticipantOutcomes::Create` service
Port the service to create a `ParticipantOutcome` from ECF, replicating the
functionality.

The `course_for_participant` validator has been omitted as its redundant; we
require a completed declaration for the given `course_identifier` to submit the
outcome which implies the participant has an application with the relevant
course.

Add transient `course` to declarations factory to more easily create a
declaration with a given course.