Account for invalid dates in a valid format
We validate the format of a date, but this still lets invalid dates through if
they are in a valid format (according to the regex).

Instead of trying to make the regex more specific with values for the day,
month and year I've opted to do a `Date.parse` validation, which should be
more robust and readable.