Add custom error message matcher
We want to be able to easily match active model errors on a given attribute,
with a certain translation key and message.

Add error messages to shoulda validation matchers (we can't easily match on
the translation key here as well as the message, so we're just matching on
the message).