Add missing answer to RegistrationWizardSpec
For some reason after adding in the custom I18n backend this test was
raising a missing translation error due to the store not having the
`referred_by_return_to_teaching_adviser` value. The value is used when
retrieving the translation and if empty was throwing an error. I'm not sure
how this worked before, but presumably its a state the application never
gets into.