Add create participant outcomes endpoint/service (#1498)
* Add `ParticipantOutcomes::Create` service

Port the service to create a `ParticipantOutcome` from ECF, replicating the
functionality.

The `course_for_participant` validator has been omitted as its redundant; we
require a completed declaration for the given `course_identifier` to submit the
outcome which implies the participant has an application with the relevant
course.

Add transient `course` to declarations factory to more easily create a
declaration with a given course.

* Remove `presence` validation from `ecf_id`

We generate this at the database level, but if we don't omit the validation
here it will prevent us from creating a new `ParticipantOutcome` without
specifying an `ecf_id` upfront.

* Add POST `api/:version/npq/participants/:id/outcomes` endpoints

Add the endpoints to create a `ParticipantOutcome` for a given participant.
Calls the `ParticipantOutcomes::Create` service.

* Fix `ParticipantOutcome::Create` permitted courses validation

We should be excluding two course identifiers from the global list.

* Add error translations for participant/lead_provider

We don't surface these to lead providers as the service won't be called if the
lead provider/participant are not present, but worth adding for completeness.

* Account for invalid dates in a valid format

We validate the format of a date, but this still lets invalid dates through if
they are in a valid format (according to the regex).

Instead of trying to make the regex more specific with values for the day,
month and year I've opted to do a `Date.parse` validation, which should be
more robust and readable.

* Remove unused translation and add presence validations

* Move queries to get latest participant outcome into scopes

Move the query to get declarations eligible for outcomes and the latest
participant outcome into scopes on the `Declaration` model so that we can
reuse them.

* Add custom i18n backend

We want to extend the default Rails i18n backend so that we can expose the
default `attribute` translation value for active model errors as a
parameterized string. This means we can easily create translations that
reference the attribute name, for example `#/completion_date`.

* Add custom error message matcher

We want to be able to easily match active model errors on a given attribute,
with a certain translation key and message.

Add error messages to shoulda validation matchers (we can't easily match on
the translation key here as well as the message, so we're just matching on
the message).

* Add missing answer to RegistrationWizardSpec

For some reason after adding in the custom I18n backend this test was
raising a missing translation error due to the store not having the
`referred_by_return_to_teaching_adviser` value. The value is used when
retrieving the translation and if empty was throwing an error. I'm not sure
how this worked before, but presumably its a state the application never
gets into.