Add X-With-Server-Date header support
In ECF we allow the server-time to be changed per-request by specifying an
`X-With-Server-Header` header.

We want to replicate this functionality in NPQ for all environments apart from
production.

Add `time_traveler` middleware from ECF and update it to use
`ActiveSupport::Testing::TimeHelpers` instead of `TimeCop`. Load middleware
into all environments, but disable in production.