Add `bullet` gem (#1527)
* Add bullet gem

Add `bullet` to identify and report N+1 queries. Enabled in `development` and
will raise an error in `test` if an N+1 is detected.

* Precompile assets using prod environment

We bundle the production assets and run the precompile using the test env,
which doesn't make sense. It causes an issue now that we have `bullet`
referenced in the `test.rb` as it installed installed with the production
gems.

Switch to compiling assets using the prod environment.