Add validation to ensure application cohort matches schedule cohort
When an application is accepted we attach a schedule. The application has a
cohort as does the schedule, so we want to validate that an application can
only have a schedule with the same cohort.

The `Accept` service ensures the schedule is picked using the application
cohort, so this will always be the case when accepting an application; its
more to protect against manual data changes.