Clean up service error tests and translations (#1535)
* Clean up service error tests and translations

Update the API service class specs to use the `have_error` matcher.

Clean up the translations file to make it consistent.

* Add validation to ensure application cohort matches schedule cohort

When an application is accepted we attach a schedule. The application has a
cohort as does the schedule, so we want to validate that an application can
only have a schedule with the same cohort.

The `Accept` service ensures the schedule is picked using the application
cohort, so this will always be the case when accepting an application; its
more to protect against manual data changes.

* Fix typo in error attribute

* Fix applications_spec flakiness

The `cohort` of the application needs to be explicitly set to the cohort of
the schedule.