Fix API sorting
The API sorting was not working correctly due to the `sort` parameter not being
passed into the underlying query objects.

Update `index` actions for applications/participants so that the `sort`
parameter tranfers to the underlying query.

Add test coverage to avoid regressions in the future.

Remove unused `sort` attribute from outcomes query.