Fix query scopes for `show` actions (#1532)
We were using the same `Query` instance for both the `index` and `show`
actions; this meant the filtering was being applied to the `show` endpoints
when it shouldn't be (although this would only happen if a filter parameter was
passed in).

Extract the query `conditions` so that we only surface them in the `index`
actions.