Change participant errors to participant_id
On various services we accept a `participant` instead of `participant_id`; this
results in the error messages being on the `participant` attribute instead of
the `participant_id` attribute supplied to the controller (as we retrieve the
participant in the controller).

We could override the paramenter name in the error messages, but a more
intuitive way feels like updating the service to accept the `participant_id`
instead of `participant` and then it will all just work as expected. This does,
however, result in quite a few changes.