Treat blank filters as if they are not specified (#1556)
Currently, if we add a blank filter to a query, for example
`?filter[cohort_start_years]=` it will return no results. Instead, we want to
treat the filter as if it wasn't specified, which is how ECF functions.

Add `FilterIgnorable` concern to centralise the logic around filtering. Treat
empty values as ignored as well as the `:ignore` symbol. Retain `nil` as a
valid/filterable value.