Fix error inconsistencies with ECF (#1560)
* Remove `invalid_schedule` error

We only want one `invalid_schedule` error for `ChangeSchedule`, which is the
same one we use for `blank`.

* Change participant errors to participant_id

On various services we accept a `participant` instead of `participant_id`; this
results in the error messages being on the `participant` attribute instead of
the `participant_id` attribute supplied to the controller (as we retrieve the
participant in the controller).

We could override the paramenter name in the error messages, but a more
intuitive way feels like updating the service to accept the `participant_id`
instead of `participant` and then it will all just work as expected. This does,
however, result in quite a few changes.

* Use `ecf_id` instead of `id`

We use `participant_id` to refer to the `ecf_id` in other places; updating to be consistent.