Various Swagger doc fixes (#1567)
* Fix funded_place description/add missing funded_place

Update the `funded_place` description to match ECF and add the missing `funded_place` attribute to the enrolments schema.

Make `funded_place` `null` in the example as the application is `pending`.

* Update change-funded-place response example

We should return the application as `accepted` in this response and
`funded_place` can be `true`.

* Make UnprocessableEntityResponse schema generic

To avoid confusion, we are changing the `course_identifier` error (which
doesn't apply in a lot of cases) to a clearly generic error.

* Fix enrolled typo in swagger docs

* Restrict sorting swagger docs to endpoints that support it

Only the applications/participants endpoints support sorting in v3.

* Add missing data attribute to declaration request

The request body should be nested in a `data` attribute.

* Add request filters to v1/2 declarations CSV endpoint

* Set clawback_statement_id to nil in v3 void response

* Add `voided` state to `ParticipantOutcome` response schema

* Fix error on collapsing participant outcomes swagger accordion

We get an error as there is a filter defined with a `nil` schema reference; we
shouldn't be defining the filter at all in this scenario.

* Fix completion_date format in swagger docs

* Fix ordering of EnroilmentCsv schema

Move `funded_place` to below `school_urn` to match the serializer.

* Set `funded_place` to `true` in accepted application response example