Parallelise test suite (#1578)
* Add Knapsack gem

This gem will let us generate the time it takes to run each spec file so that
the test runner can more efficiently divide the tests between workers.

* Parallelise test suite in CI

Generate the knapsack spec report and update the GitHub CI to run a matrix for
the test suite.

* Remove unused spec file

This spec is empty; it includes a context but has no actual tests.