Change Declaration#declaration_date to DateTime (#1583)
* Change Declaration#declaration_date to DateTime

The `Declaration#declaration_date` is a `Date` in NPQ reg and a `DateTime` in
ECF. The side-effect of having it as a `Date` in NPQ reg is that the `rfc3889`
format uses an explicit timezone offset of `+00:00` instead of the shorthand
`Z` that is used for `DateTime`. We're concerned this might break provider
integrations, so to keep it consistently using the `Z` shorthand we are
changing the field to a `DateTime` in NPQ reg.

* Update declaration_date in swagger docs

Update to match the new formatting.