Send coverage/test reports to Sonarcloud
We want to send the test coverage and reports to Sonarcloud so that we can get
details on the test coverage in pull requests.

Move linting into rspec action so that we can also send the Rubocop output to Sonarcloud.

The whole setup is complicated quite a lot by the testing matrix; we need to retain the coverage/test report assets from each worker, which we then combine before sending to Sonarcloud.