Change `Schedule#allowed_declaration_types` to `declaration_types` enum (#1601)
We have duplicate enum types for declaration type; we want to use the
`declaration_type` on `Schedule` and remove the `schedule_declaration_types`
enum.

This turned out to be non-trivial due to Postgres not supporting changing the
type of an enum/array column. We have to remove the default and cast all the
values as part of the migration.

As we're only running in sandbox this is probably fine, but it has the
potential for data issues during this migration (the default will be removed
while the casting takes place, so could introduce bad data). We can write a
more robust/complex migration to mitigate this, but its probably not worth it
for sandbox and the chances are it will run without issue.