Improve performance of data migration
This commit is a big overhaul of how the migration runs. Previously, the migrators would run in sequence using a single worker to iterate through the different types of models to be migrated. This resulted in one, long-running job that consumed a lot of resources and often crashed/failed.

I thought about a few different ways of optimising this; but the route I took was to allow splitting the migrators so that they could be ran across multiple jobs, hopefully improving the throughput. This had quite a few challenges around coordinating the workers in order to avoid race conditions and ensuring the migrators ran in a suitable order (the statements migrator, for example, must run after the lead providers and cohorts have been migrated).

With this change, the `MigrationJob` kicks everything off, running the `Coordinator`, which will find the next runnable migrator (where its dependencies have migrated and it hasn't already been queued). We then kick off multiple `MigratorJob` workers which process a range of the models. When all models have been processed we re-queue the `MigrationJob` to kick off the next runnable migration.

The migrators are ran in sequence to avoid race conditions (if two migrators finish and both kick off `MigratorJob` instances we could end up with multiple migrations being coordinated for the same model). A new `queued_at` attribute has also been added to help avoid race conditions. We create a `DataMigration` for each worker and model (I tried keeping it to one `DataMigration` per model but this also had concurrency issues). The last worker to finish (for a particular model) will kick off the migration for the next model. Jobs are queued as 'high priority' to ensure they are prioritised over anything else going on.

In some cases we want to run some migration-specific code only once per model migration; a `run_once` helper is provided for this (used to log applications in NPQ not in ECF as failures, currently).

I also made various low-level performance improvements to a couple of the migrators to hopefully speed them up.

The delayed job web interface has also been fixed so super admins can access it to keep an eye on the jobs.

Areas that I want to address in a follow-up change:

- Test coverage on the individual migrators could be better (application/statement migrator attributes being set, find/create behaviour in general)
- We are using exceptions for flow control; I want to move away from that
- We are DB-bound now (on CPU); I want to bump to a bigger instance to see the effect it has on the runtime. We're constrained to 5 workers at the moment, which keeps the DB CPU load around 95%.
- We could hook into Grafana to get better metrics on how things are migrating
- I need to think about if the whole process is atomic/what happens if a migrator fails and retries
- The structure of the migration/migrators changed a lot, I want to do some tidying up on the migrator base class at some point