Add state machine to `Declaration` model (#1600)
* Add state machine to Declaration model

We want to add a state machine to the `Declaration` model to be
consistent with `Statement` and `StatementItem`.

Add state machine, defining transitions.

* Use state machine where possible

We can clean up/remove some of our methods to use the convenience methods provided by the state machine library.

* Only allow submitted -> ineligible transition

We only allow submitted -> ineligible transitions in ECF, not from billable
states.