Add Schedule migrator
Add the `Schedule` migrator to bring across all NPQ schedules from ECF. As part
of this change we also now create `CourseGroup` during the `Course` migration
(we can't do this in its own migrator as it doesn't fir the pattern of bringing
across records from ECF).

A `Schedule` must match on cohort, name, identifier and course group (derived
from the ECF schedule type). We set the applies from/to based on the milestone
dates. All milestone dates for ECF schedules must be the same or a failure is
recorded.