Migrate `ParticipantIdChange` records from ECF (#1642)
* Add ParticipantIdChange migrator

We want to migrate participant ID changes from ECF to NPQ reg. We have to do
this in a separate migrator (to the users migrator) as we need to migrate the
users first (the `ParticipantIdChange` model does not store the `ecf_id`,
instead using the NPQ reg ids).

* Migrate created_at on ParticipantIdChange records

As we expose this in the serializers we want to keep it consistent with ECF.