Clear migration failures prior to running
As we store failures in Redis, when we clear the NPQ reg database and re-run a
migration after a failed migration the failure manager will add the failures to
those from the last run (as the `DataMigration` models/ids get reset so are the
same as the previous run).

To avoid this, we can purge the failures from Redis for a `DataMigration` as
part of the migration preparation.

The `FailureManager` is becoming a not so great mix of instance/class methods;
it could do with a tidy up but its probably not worth the effort given this
code is going to be relatively short-lived.