Migrate `Declarations` from ECF (#1649)
Migrate `Declaration` records from ECF. We run a second migrator to set the
`superseded_by_id` on the migrated declarations, as we need to be able to look
up their NPQ reg ids for this.

We match declarations on `ecf_id` and update the attributes with the values in
ECF. We may need to optimise the application lookup depending on runtime.