Migrate ECF applications to NPQ reg
We want to bring across the data associated with NPQ applications in ECF to NPQ
reg. Previously we were performing an integrity check on the attributes, but
after investigation we determined all attributes should be updated from ECF to
NPQ reg for applications.

The school, course, lead provider and user should not be different and we have
added an integrity check to ensure this.