Fix attribute slicing in migrators
In the `Contract` and `ApplicationState` migrators we slice attributes using an
array; this results in an empty hash and the tests were passing as we made the
same slicing error. We need to use the splat operator here to correctly slice
the attributes. In the `Contract` migrator we were also using symbols in the
tests instead of strings, which is why the test was also incorrectly passing
(slicing by an attribute that doesn't exist omits it).