Fix bugs in Contract migrator (#1697)
* Optimise lookups in Contract migrator

To be consistent with other migrators and to hopefully speed things up we
preload the lookups (course/statement).

* Fix attribute slicing in migrators

In the `Contract` and `ApplicationState` migrators we slice attributes using an
array; this results in an empty hash and the tests were passing as we made the
same slicing error. We need to use the splat operator here to correctly slice
the attributes. In the `Contract` migrator we were also using symbols in the
tests instead of strings, which is why the test was also incorrectly passing
(slicing by an attribute that doesn't exist omits it).