Add ecf_id to remaining migration tables
Add the `ecf_id` to the remaining migration tables, apart from `schools` as
these already exist in NPQ reg and are only verified/not actually migrated as
part of the migration process.

The `ecf_id` fields are nullable since we don't expose them to lead providers
and will not need to generate them going forward.