Add callback to schedule factory
Setting the schedule dates to 1 week ago is to ensure that
without time travel declaration factories are valid. Declaration model
validations require the `declaration_date` not to be in the future, and
to be after the application schedule `applies_from` date. We can disable
this callback via the transient `change_applies_dates` attribute by
setting it to false.

We are disabling schedule dates in seed data, as well as setting up
production like data in the shared context default schedules.