Add `disabled_at` to provider models and exclude by default
We want to include `IttProvider` and `PrivateChildCareProvider` instances from
ECF that are not currently in NPQ reg. In order to bring them across as part of
the migration process and not change the current behaviour in NPQ reg, we will
flag any new providers as disabled as part of the migration.

A `default_scope` has been added to ensure that existing behaviour will
continue to only show the providers currently in NPQ reg (and not disabled).