Migrate IttProvider and PrivateChildcareProvider
We want to bring across any IttProvider or PrivateChildcareProvider records
that exist in ECF (as part of an NPQApplication) but are not present in NPQ
reg. When we bring them across we want to flag them as disabled so that they
don't appear in other areas of the NPQ reg application.

Providers are matched case insensitively. We can also fallback on the
`operating_name` of the `IttProvider` if set.