Workaround for declaration validation failures in specs (#1711)
* Revert "Fix declaration_date mismatch with schedule (#1654)"

This reverts commit 2dd70407ba9fc6b8127e61151478d7c9846bba06.

* Add callback to schedule factory

Setting the schedule dates to 1 week ago is to ensure that
without time travel declaration factories are valid. Declaration model
validations require the `declaration_date` not to be in the future, and
to be after the application schedule `applies_from` date. We can disable
this callback via the transient `change_applies_dates` attribute by
setting it to false.

We are disabling schedule dates in seed data, as well as setting up
production like data in the shared context default schedules.