Fix Codespaces not booting (#1726)
Codespaces fails to run since we changed the asset pipeline; we need to run
`yarn build` and `yarn build:css` to compile the JS/CSS assets. Instead of
doing this once on boot I've opted to run Foreman in the Codespace, which gives
us the `--watch` behaviour to recompile CSS/JS as and when it changes.

The codespace was also failing to boot due to `bundle install` failing to
compile native gems (`bigdecimal`). I went on a bit of a wild goose chase with
this error; initially I thought it was because the `docker-compose.yml`
recently changed, but Codespaces uses a different one. I then thought it was
due to native libraries not being present so tried installing various
dependencies (`build-essential`, `ruby-dev`, etc). None of this worked; in the
end it turned out to be a recent upgrade to Ruby breaking `bigdecimal` on
Ubuntu. The fix is to install `clang`, which requires the `universe`
repository.