Add `ecf_id` to remaning migration tables (#1708)
* Add ecf_id to remaining migration tables

Add the `ecf_id` to the remaining migration tables, apart from `schools` as
these already exist in NPQ reg and are only verified/not actually migrated as
part of the migration process.

The `ecf_id` fields are nullable since we don't expose them to lead providers
and will not need to generate them going forward.

* Use `ecf_id` in all migrators

We now have `ecf_id` on cohorts, contracts, participant_id_changes, schedules
and statement_items, so we want to match on this attribute going forward as
part of the  migration process.

* Optimise cohort lookups with ecf_id

Use `ecf_id` for cohort lookups to avoid including the cohort in the underlying
queries.

* Clean up application migrator to use schedule ecf_id

As we now have `ecf_id` on schedule we can perform a simpler lookup.

* Revert adding `ecf_id` on `contracts` table

There isn't a 1:1 mapping here; ECF contracts already get tied back to
`ContractTemplate` and the `Contract` is linked to a `ContractTemplate`, so we
can infer it that way already.

* Fix failing test after merging main

* Fix flakey cohort spec

* Fix flakey separation shared data spec