Create IttProvider and PrivateChildcareProvider on migration (#1716)
* Migrate IttProvider and PrivateChildcareProvider

We want to bring across any IttProvider or PrivateChildcareProvider records
that exist in ECF (as part of an NPQApplication) but are not present in NPQ
reg. When we bring them across we want to flag them as disabled so that they
don't appear in other areas of the NPQ reg application.

Providers are matched case insensitively. We can also fallback on the
`operating_name` of the `IttProvider` if set.

* Add test to ensure disabled providers are serialized

As we have added a `default_scope` to the `IttProvider` and
`PrivateChildcareProvider` to ignore those that are disabled, we want to add an
explicit test to ensure an application with a disbaled provider is still
correctly serialized.