Only raise on multiple application TRNs if no teacher profile TRN (#1729)
We only want to raise an error for multiple verified `NPQApplication` TRNs if
there is not a `teacher_profile.trn` on the ECF user. We will favour the
profile TRN over the `NPQApplication` TRNs and fallback on the NPQ user TRN in
case there is not a verified TRN in ECF (and the NPQ user already has one).