Cache migration dependencies in Redis
When a migrator is queued we will now pre-warm the dependency cache in Redis.
We cache the lookup hash as individual key/value entries in Redis so we don't
have to load the whole lot into memory for each worker.

As we are storing lots of individual keys in Redis, clearing them manually
becomes difficult. Instead, I opted to set a 1 hour TTL on the cache keys,
which should be long enough for any single run but short enough that the cache
is cleared before the next run. That being said, even if the cache is still
warm from a previous run it only stores ID mappings which are unlikely to
change between runs, so it shouldn't cause us issues.