Add GitHub action to report nightly migration status
Add a GitHub action that will run at 8am every day and report the outcome of
the nightly data migration.

If there are no failures a success message is posted to Slack (mainly so we
know its running successfully/daily). If there are failures, a message is
posted to Slack containing the details of the workers/models that failed and
the percentage that were migrated successfully.