Fix failing specs due to declaration date (#1759)
A number of tests setup the declaration_date to be `1.day.ago`; this is usually
fine as the `applies_from` of the schedule is the start of the month; but when
we're on the first day of the next month it causes the date to be too
early/before the `applies_from`.

As a workaround we can go back `1.hour` instead of `1.day`, so the date is on
the 1st of the month and not the previous month.

This gets the tests passing; we may want to find a better solution as they
would still fail when ran between 12-1am on the 1st of a month.