Run migration on a nightly schedule (#1752)
* Refresh migration DB nightly

We want to refresh the migration DB every night so that we can run themigration on a schedule.

Refresh migration DB at midnight.

* Run migration nightly

We want to run the data migration on a nightly basis to monitor how it
performs/the results.

Run migration at 12:30am every morning; this is 30 minutes after the migration
DBs are refreshed.

* Add GitHub action to report nightly migration status

Add a GitHub action that will run at 8am every day and report the outcome of
the nightly data migration.

If there are no failures a success message is posted to Slack (mainly so we
know its running successfully/daily). If there are failures, a message is
posted to Slack containing the details of the workers/models that failed and
the percentage that were migrated successfully.

* Disable migration in staging/review

As we are now running a cron off the back of this flag, we only want it enabled in development and the migration environment.