Schedule cron jobs on db rake tasks
Currently, if we add a new cron job we need to remember to manually schedule it
the first time or it won't ever run. To avoid this, the gem recommends hooking
into the `db:migrate` and `db:schema:load` rake tasks to also run the cron
scheduling (it will ignore jobs already scheduled and ensure new jobs are
scheduled).

Set to run only on hosted environments. I manually ran the rake task in the
migration environment to ensure it works and to get the migration cron job
scheduled.